update files for 11ty upgrade and update img shortcodes
