+++ src/_data/project.js
@@ -1,3 +1,3 @@
-module.exports = {
+export default {
   environment: process.env.ELEVENTY_ENV
 };
