+++ src/_includes/partials/footer.njk
@@ -5,8 +5,8 @@
     </div>
     <div>
       <div class="grid grid-flow-col gap-4">
-        <a class="" target="_blank" href="https://github.com/zapalblizh">{% svg "github-logo" %}</a>
-        <a class="" target="_blank" href="https://www.linkedin.com/in/aleksandr-zapparov/">{% svg "linkedin-logo" %}</a>
+        <a class="" target="_blank" href="https://github.com/zapalblizh">{% svg "github-logo", "size-7" %}</a>
+        <a class="" target="_blank" href="https://www.linkedin.com/in/aleksandr-zapparov/">{% svg "linkedin-logo", "size-7" %}</a>
       </div>
     </div>
   </div>
