+++ src/content/index.njk
@@ -19,8 +19,8 @@ permalink: /
       <p>In my free time, I love to play chess, solve puzzles, play video games and gain new knowledge and news about technological news.</p>
     </div>
 
-    {% image "../assets/img/zapalblizh.jpeg", "Picture of Aleksandr Zapparov", [360],
-      "mx-auto rounded-md" %}
+    <img class="mx-auto rounded-md" src="../assets/img/zapalblizh.jpeg" alt="Picture of Aleksandr Zapparov" />
+
   </div>
 
 </div>
